recipe_hub/backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
Pillow
supabase==2.11.0
PyJWT==2.8.0
psycopg2-binary==2.9.11
dj-database-url==2.1.0